+++ service/languageServer/src/test/java/science/aist/sushiya/service/languageserver/completion/AliasCompletionProviderTest.java
@@ -1,32 +1,32 @@
-package science.aist.sushiya.service.languageserver;
+package science.aist.sushiya.service.languageserver.completion;
 
 import org.eclipse.lsp4j.*;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import science.aist.sushiya.service.languageserver.completion.FHIRResources;
-import science.aist.sushiya.service.languageserver.completion.InstanceOfCompletionProvider;
+import science.aist.sushiya.service.languageserver.FSHFileHandler;
+import science.aist.sushiya.service.languageserver.completion.AliasCompletionProvider;
 
 /**
  * <p>Created by Sophie Bauernfeind on 25.03.2021</p>
- * <p>Test class for {@link InstanceOfCompletionProvider}</p>
+ * <p>Test class for {@link AliasCompletionProvider}</p>
  *
  * @author Sophie Bauernfeind
  */
-public class InstanceOfCompletionProviderTest {
-    private static final InstanceOfCompletionProvider provider = new InstanceOfCompletionProvider();
+public class AliasCompletionProviderTest {
+    private static final AliasCompletionProvider provider = new AliasCompletionProvider();
     private static final String uri = "testing";
 
     @BeforeMethod
     public void setUp() {
         FSHFileHandler.getInstance().clean();
     }
-    
+
     @Test
     public void testActivation1(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -47,7 +47,7 @@ public class InstanceOfCompletionProviderTest {
     public void testActivation2(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "  InstanceOf: ";
+        String text = "  Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -68,7 +68,7 @@ public class InstanceOfCompletionProviderTest {
     public void testActivation3(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "  InstanceOf  : ";
+        String text = "  Alias  : ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -86,10 +86,32 @@ public class InstanceOfCompletionProviderTest {
     }
 
     @Test
+    public void testActivation4(){
+        //given
+        TextDocumentItem textDocumentItem = new TextDocumentItem();
+        String text = "Alias: LNC = http://loinc.org\n"
+                + "  Alias  : ";
+        textDocumentItem.setText(text);
+        textDocumentItem.setUri(uri);
+
+        CompletionParams params = new CompletionParams();
+        Position position = new Position(1,11);
+        params.setPosition(position);
+        CompletionContext completionContext = new CompletionContext();
+        completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
+        completionContext.setTriggerCharacter(" ");
+        params.setContext(completionContext);
+        //when
+
+        //then
+        Assert.assertTrue(provider.test(textDocumentItem,params));
+    }
+
+    @Test
     public void testNoActivationWrongPosition() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -110,7 +132,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationOutOfBound() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -152,7 +174,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationIncorrectText() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "test InstanceOf: ";
+        String text = "test Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -173,7 +195,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationNoSetPosition() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -192,7 +214,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationNoSetContext() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -209,7 +231,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationNoSetTriggerKind() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -229,7 +251,7 @@ public class InstanceOfCompletionProviderTest {
     public void testNoActivationNoSetTriggerCharacter() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -249,7 +271,7 @@ public class InstanceOfCompletionProviderTest {
     public void testActivationNoSetUri(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: ";
+        String text = "Alias: ";
         textDocumentItem.setText(text);
 
         CompletionParams params = new CompletionParams();
@@ -270,9 +292,8 @@ public class InstanceOfCompletionProviderTest {
     public void testAmountCompletionItems(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: "
-                + "\n"
-                + "Profile: Test";
+        String text = "Alias: test = testing \n"
+                    + "Alias: test2 = testing\n";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -283,6 +304,7 @@ public class InstanceOfCompletionProviderTest {
         completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
         completionContext.setTriggerCharacter(" ");
         params.setContext(completionContext);
+
         //when
         DidOpenTextDocumentParams openParams = new DidOpenTextDocumentParams();
         openParams.setTextDocument(textDocumentItem);
@@ -290,21 +312,15 @@ public class InstanceOfCompletionProviderTest {
         provider.test(textDocumentItem,params);
 
         //then
-        Assert.assertEquals(provider.get().size(),
-                FHIRResources.getInstance().getAllBase().size()+
-                        FHIRResources.getInstance().getAllClinical().size() +
-                        1);
+        Assert.assertEquals(provider.get().size(),4);
     }
 
     @Test
     public void testAmountCompletionItems2(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "InstanceOf: \n"
-                + "\n"
-                + "Profile: Test\n"
-                + "\n"
-                + "Profile: Test2";
+        String text = "Profile: Testing \n"
+                    + "Alias: test = testing\n" ;
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -315,6 +331,7 @@ public class InstanceOfCompletionProviderTest {
         completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
         completionContext.setTriggerCharacter(" ");
         params.setContext(completionContext);
+
         //when
         DidOpenTextDocumentParams openParams = new DidOpenTextDocumentParams();
         openParams.setTextDocument(textDocumentItem);
@@ -322,9 +339,33 @@ public class InstanceOfCompletionProviderTest {
         provider.test(textDocumentItem,params);
 
         //then
-        Assert.assertEquals(provider.get().size(),
-                FHIRResources.getInstance().getAllBase().size()+
-                        FHIRResources.getInstance().getAllClinical().size() +
-                        2);
+        Assert.assertEquals(provider.get().size(),3);
+    }
+
+    @Test
+    public void testAmountCompletionItems3(){
+        //given
+        TextDocumentItem textDocumentItem = new TextDocumentItem();
+        String text = " Alias: test = testing\n"
+                    + "Alias  : test2 = testing\n";
+        textDocumentItem.setText(text);
+        textDocumentItem.setUri(uri);
+
+        CompletionParams params = new CompletionParams();
+        Position position = new Position(0,text.length());
+        params.setPosition(position);
+        CompletionContext completionContext = new CompletionContext();
+        completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
+        completionContext.setTriggerCharacter(" ");
+        params.setContext(completionContext);
+
+        //when
+        DidOpenTextDocumentParams openParams = new DidOpenTextDocumentParams();
+        openParams.setTextDocument(textDocumentItem);
+        FSHFileHandler.getInstance().addFile(openParams);
+        provider.test(textDocumentItem,params);
+
+        //then
+        Assert.assertEquals(provider.get().size(),4);
     }
 }
