+++ service/languageServer/src/test/java/science/aist/sushiya/service/languageserver/completion/CompletionProcessorTest.java
@@ -1,4 +1,4 @@
-package science.aist.sushiya.service.languageserver;
+package science.aist.sushiya.service.languageserver.completion;
 
 import org.eclipse.lsp4j.*;
 import org.testng.Assert;
