+++ service/languageServer/src/test/java/science/aist/sushiya/service/languageserver/completion/FHIRResourcesTest.java
@@ -1,8 +1,7 @@
-package science.aist.sushiya.service.languageserver;
+package science.aist.sushiya.service.languageserver.completion;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import science.aist.sushiya.service.languageserver.completion.FHIRResources;
 
 /**
  * <p>Created by Sophie Bauernfeind on 26.03.2021</p>
