+++ service/languageServer/src/test/java/science/aist/sushiya/service/languageserver/completion/SourceCompletionProviderTest.java
@@ -1,19 +1,20 @@
-package science.aist.sushiya.service.languageserver;
+package science.aist.sushiya.service.languageserver.completion;
 
 import org.eclipse.lsp4j.*;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import science.aist.sushiya.service.languageserver.completion.AliasCompletionProvider;
+import science.aist.sushiya.service.languageserver.FSHFileHandler;
+import science.aist.sushiya.service.languageserver.completion.SourceCompletionProvider;
 
 /**
  * <p>Created by Sophie Bauernfeind on 25.03.2021</p>
- * <p>Test class for {@link AliasCompletionProvider}</p>
+ * <p>Test class for {@link SourceCompletionProvider}</p>
  *
  * @author Sophie Bauernfeind
  */
-public class AliasCompletionProviderTest {
-    private static final AliasCompletionProvider provider = new AliasCompletionProvider();
+public class SourceCompletionProviderTest {
+    private static final SourceCompletionProvider provider = new SourceCompletionProvider();
     private static final String uri = "testing";
 
     @BeforeMethod
@@ -25,7 +26,7 @@ public class AliasCompletionProviderTest {
     public void testActivation1(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -46,7 +47,7 @@ public class AliasCompletionProviderTest {
     public void testActivation2(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "  Alias: ";
+        String text = "  Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -67,7 +68,7 @@ public class AliasCompletionProviderTest {
     public void testActivation3(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "  Alias  : ";
+        String text = "  Source  : ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -85,32 +86,10 @@ public class AliasCompletionProviderTest {
     }
 
     @Test
-    public void testActivation4(){
-        //given
-        TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: LNC = http://loinc.org\n"
-                + "  Alias  : ";
-        textDocumentItem.setText(text);
-        textDocumentItem.setUri(uri);
-
-        CompletionParams params = new CompletionParams();
-        Position position = new Position(1,11);
-        params.setPosition(position);
-        CompletionContext completionContext = new CompletionContext();
-        completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
-        completionContext.setTriggerCharacter(" ");
-        params.setContext(completionContext);
-        //when
-
-        //then
-        Assert.assertTrue(provider.test(textDocumentItem,params));
-    }
-
-    @Test
     public void testNoActivationWrongPosition() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -131,7 +110,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationOutOfBound() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -173,7 +152,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationIncorrectText() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "test Alias: ";
+        String text = "test Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -194,7 +173,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationNoSetPosition() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -213,7 +192,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationNoSetContext() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -230,7 +209,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationNoSetTriggerKind() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -250,7 +229,7 @@ public class AliasCompletionProviderTest {
     public void testNoActivationNoSetTriggerCharacter() {
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -270,7 +249,7 @@ public class AliasCompletionProviderTest {
     public void testActivationNoSetUri(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: ";
+        String text = "Source: ";
         textDocumentItem.setText(text);
 
         CompletionParams params = new CompletionParams();
@@ -291,8 +270,9 @@ public class AliasCompletionProviderTest {
     public void testAmountCompletionItems(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Alias: test = testing \n"
-                    + "Alias: test2 = testing\n";
+        String text = "Source: \n"
+                + "\n"
+                + "Profile: Test";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -311,15 +291,18 @@ public class AliasCompletionProviderTest {
         provider.test(textDocumentItem,params);
 
         //then
-        Assert.assertEquals(provider.get().size(),4);
+        Assert.assertEquals(provider.get().size(),1);
     }
 
     @Test
     public void testAmountCompletionItems2(){
         //given
         TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = "Profile: Testing \n"
-                    + "Alias: test = testing\n" ;
+        String text = "Source: \n"
+                + "\n"
+                + "Profile: Test"
+                + "\n"
+                + "Profile: Test2";
         textDocumentItem.setText(text);
         textDocumentItem.setUri(uri);
 
@@ -338,33 +321,6 @@ public class AliasCompletionProviderTest {
         provider.test(textDocumentItem,params);
 
         //then
-        Assert.assertEquals(provider.get().size(),3);
-    }
-
-    @Test
-    public void testAmountCompletionItems3(){
-        //given
-        TextDocumentItem textDocumentItem = new TextDocumentItem();
-        String text = " Alias: test = testing\n"
-                    + "Alias  : test2 = testing\n";
-        textDocumentItem.setText(text);
-        textDocumentItem.setUri(uri);
-
-        CompletionParams params = new CompletionParams();
-        Position position = new Position(0,text.length());
-        params.setPosition(position);
-        CompletionContext completionContext = new CompletionContext();
-        completionContext.setTriggerKind(CompletionTriggerKind.TriggerCharacter);
-        completionContext.setTriggerCharacter(" ");
-        params.setContext(completionContext);
-
-        //when
-        DidOpenTextDocumentParams openParams = new DidOpenTextDocumentParams();
-        openParams.setTextDocument(textDocumentItem);
-        FSHFileHandler.getInstance().addFile(openParams);
-        provider.test(textDocumentItem,params);
-
-        //then
-        Assert.assertEquals(provider.get().size(),4);
+        Assert.assertEquals(provider.get().size(),2);
     }
 }
