+++ service/languageServer/src/test/java/science/aist/sushiya/service/languageserver/completion/VsRuleCompletionProviderTest.java
@@ -1,9 +1,10 @@
-package science.aist.sushiya.service.languageserver;
+package science.aist.sushiya.service.languageserver.completion;
 
 import org.eclipse.lsp4j.*;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+import science.aist.sushiya.service.languageserver.FSHFileHandler;
 import science.aist.sushiya.service.languageserver.completion.VsRuleCompletionProvider;
 
 /**
